feat: 사용자 페이지 왼쪽 사이드바 1차 렌더링 (#425)
