+++ src/app/userpage/(template)/layout.tsx
@@ -1,4 +1,4 @@
-import UserPagePrevButton from "@components/Layout/UserPagePrevButton";
+import UserPageLeftSidebar from "@components/Layout/UserPageLeftSideBar";
 export const metadata = {
   title: "Componique",
   description: "Generated by Next.js",
@@ -11,8 +11,8 @@ export default function DocsLayout({
 }) {
   return (
     <div className="w-full">
-      <section className="fixed left-3 top-28 h-[calc(100%-120px)]">
-        <UserPagePrevButton />
+      <section className="fixed left-4 top-26 h-[calc(100%-120px)] w-[230px] rounded-l-2xl bg-[#D8EAF8] shadow-lg dark:bg-Navy">
+        <UserPageLeftSidebar />
       </section>
       <div className="w-full">{children}</div>
     </div>
